Refactor non-Qt functionality of export dialog into a base class so that we can reuse it for a Jupyter dialog.
